# src/config/swagger.yaml
swagger: "2.0"
info:
  description: "GestIC App API documentation"
  version: "1.0.0"
  title: "GestIC API"
host: "localhost:3333"
...
paths:
...
  /project:
    get:
      tags:
      - "Project"
      summary: "Get a list of all active projects"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Obtained a list of all active projects"
          schema:
            type: "array"
            produces: "application/json"
            items:
              type: "object"
              properties:
                id:
                  type: "string"
                  format: "uuid"
                userId:
                  type: "string"
                  format: "uuid"
                name:
                  type: "string"
                description:
                  type: "string"
        "500":
          description: "Could not obtain a list of all active projects"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    post:
      tags:
      - "Project"
      summary: "Insert a project"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "object"
            properties:
              name:
                type: "string"
              description:
                type: "string"
              userId:
                type: "string"
                format: "uuid"
      responses:
        "200":
          description: "Inserted a project"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              userId:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
        "500":
          description: "Could not insert a project"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    put:
      tags:
      - "Project"
      summary: " Update project information"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          description: "Any quantity of the following parameters can be sent, but id is required."
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              userId:
                type: "string"
                format: "uuid"
      responses:
        "200":
          description: "Sucessfully updated a project"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              userId:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
        "500":
          description: "Could not update a project"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
  /project/:id:
    get:
      tags:
        - "Project"
      summary: "Get project information"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Obtained a project"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              userId:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
        "500":
          description: "Could not obtain a project"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    delete:
      tags:
        - "Project"
      summary: "Delete project"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Successfully removed a project"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              userId:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
        "500":
          description: "Could not remove a project"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
...
  /research:
    get:
      tags:
      - "Research groups"
      summary: "Get a list of all research groups"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Obtained a list of all active research groups"
          schema:
            type: "array"
            produces: "application/json"
            items:
              type: "object"
              properties:
                id:
                  type: "string"
                  format: "uuid"
                name:
                  type: "string"
                description:
                  type: "string"
                activities:
                  type: "string"
        "500":
          description: "Could not obtain a list of all active research groups"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    post:
      tags:
      - "Research groups"
      summary: "Insert a research group"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "object"
            properties:
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
      responses:
        "200":
          description: "Inserted a research group"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
        "500":
          description: "Could not insert a research group"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    put:
      tags:
      - "Research groups"
      summary: " Update research group information"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          description: "Any quantity of the following parameters can be sent, but id is required."
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
      responses:
        "200":
          description: "Sucessfully updated a research group"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
        "500":
          description: "Could not update a research group"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
  /research/:id:
    get:
      tags:
        - "Research groups"
      summary: "Get research group information"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Obtained a research group"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
              members:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "string"
                      format: "uuid"
                    name:
                      type: "string"
                    email:
                      type: "string"
        "500":
          description: "Could not obtain a research group"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    delete:
      tags:
        - "Research groups"
      summary: "Delete research group"
      produces:
      - "application/json"
      parameters:
        - in: param
          name: id
          required: true
          type: string
          format: "uuid"
      responses:
        "200":
          description: "Successfully removed a research group"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              userId:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
        "500":
          description: "Could not remove a research group"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
  /research/members:
    post:
      tags:
      - "Research groups"
      summary: "Add a member to research group"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              userId:
                type: "string"
                format: "uuid"
      responses:
        "200":
          description: "Added a member"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
              members:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "string"
                      format: "uuid"
                    name:
                      type: "string"
                    email:
                      type: "string"
        "500":
          description: "Could not add a member to research group"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
    delete:
      tags:
      - "Research groups"
      summary: "Remove a member to research group"
      produces:
      - "application/json"
      consumes:
      - "application/json"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              userId:
                type: "string"
                format: "uuid"
      responses:
        "200":
          description: "Removed a member"
          schema:
            type: "object"
            properties:
              id:
                type: "string"
                format: "uuid"
              name:
                type: "string"
              description:
                type: "string"
              activities:
                type: "string"
              members:
                type: "array"
                items:
                  type: "object"
                  properties:
                    id:
                      type: "string"
                      format: "uuid"
                    name:
                      type: "string"
                    email:
                      type: "string"
        "500":
          description: "Could not remove a member from research group"
          schema:
            type: "object"
            properties:
              message:
                type: "string"